requirements.txt
requests==2.31.0
psutil==5.9.6
watchdog==3.0.0
python-dateutil==2.8.2